rename taskexec mode to task and add more operations
