+++ src/nomad_tools/entry.py
@@ -14,7 +14,7 @@ from . import (
     entry_listattributes,
     entry_nodenametoid,
     entry_port,
-    entry_taskexec,
+    entry_task,
     entry_vardir,
     entry_watch,
 )
@@ -78,7 +78,7 @@ cli.add_command(entry_go.cli)
 cli.add_command(entry_listattributes.cli)
 cli.add_command(entry_nodenametoid.cli)
 cli.add_command(entry_port.cli)
-cli.add_command(entry_taskexec.cli)
+cli.add_command(entry_task.cli)
 cli.add_command(entry_vardir.cli)
 cli.add_command(entry_watch.cli)
